Refactor field-function implementations out of LBSProblem and DiscreteOrdinatesProblem.
